fix: move logica to authController instead of user
